hud: fractional 2D auto-scale; hud_wide defaults to classic layout
gl3state.scale becomes float and the auto mode (gl_2dscale 0) holds a
constant 600-line virtual layout at every resolution and aspect, so
the HUD keeps the same share of the screen whether windowed,
fullscreen, or stretched -- the ortho projection does the sampling, so
fractional scale is fine. Explicit gl_2dscale values (which may now be
fractional) are still honored.

hud_wide default flips 1 -> 0: the classic centered statusbar, whose
layout never changes with aspect, is the default; the edge-anchored
16:9 layout is opt-in.

+++ src/ref_gl3/gl3_local.h
@@ -41,7 +41,7 @@ typedef struct
 {
 	int			width, height;		// current drawable size in pixels
 	int			vw, vh;				// virtual 2D resolution reported to the client
-	int			scale;				// real pixels per virtual pixel (UI scale)
+	float		scale;				// real pixels per virtual pixel (UI scale)
 	qboolean	fullscreen;
 	SDL_Window	*window;
 	SDL_GLContext	context;
